logo/requirements.txt
numpy
pillow
matplotlib
